+++ packages/agents/lighthouse/src/bindings/backend/index.ts
@@ -5,6 +5,7 @@ import {
   NxtpError,
   formatUrl,
   convertFromDbTransfer,
+  transfersCastForUrl,
 } from "@connext/nxtp-utils";
 import { constants } from "ethers";
 import axios from "axios";
@@ -37,7 +38,9 @@ export const pollBackend = async () => {
   const { logger } = getContext();
   const { execute } = getOperations();
 
+  logger.debug("Polling backend", requestContext, methodContext, {});
   const reconciledTransactions = await getReconciledTransactions();
+  logger.debug("Get reconciled transactions", requestContext, methodContext, { reconciledTransactions });
 
   await Promise.all(
     reconciledTransactions.map(async (transaction: any) => {
@@ -79,14 +82,11 @@ export const pollBackend = async () => {
 
 export const getReconciledTransactions = async (): Promise<any> => {
   const { requestContext, methodContext } = createLoggingContext(getReconciledTransactions.name);
-  const {
-    // adapters: {},
-    logger,
-    config,
-  } = getContext();
+  const { logger, config } = getContext();
 
-  const statusIdentifier = `status=eq.Reconciled&`;
+  const statusIdentifier = `status=eq.Reconciled&${transfersCastForUrl}`;
   const uri = formatUrl(config.backendUrl, "transfers?", statusIdentifier);
+  logger.debug("Getting transactions from URI", requestContext, methodContext, { uri });
   try {
     const response = await axios.get(uri);
     return response.data;