+++ packages/agents/sdk/src/lib/errors/index.ts
@@ -11,3 +11,21 @@ export class ChainDataUndefined extends NxtpError {
     super("Chain Data Undefined", context, ChainDataUndefined.name);
   }
 }
+
+export class ParamsInvalid extends NxtpError {
+  constructor(context: any = {}) {
+    super("Params invalid", context, ParamsInvalid.name);
+  }
+}
+
+export class UriInvalid extends NxtpError {
+  constructor(context: any = {}) {
+    super("Uri invalid", context, UriInvalid.name);
+  }
+}
+
+export class ApiRequestFailed extends NxtpError {
+  constructor(context: any = {}) {
+    super("Api Request failed", context, ApiRequestFailed.name);
+  }
+}